Add files via upload
This program was developed to determine the intensity of a electron beam on a HAADF detector when the electron beam passes through a wedge shaped material (1nm - 3 microns). The thickness ranges from one nm to few microns. The program calculates elastic scattering from an electron beam per nm of the sample and also the absorption using scattering coefficients, density, atomic number, absorption parameters and HAADF scattering angle and detector. 
## parameters.